requirements.txt
astroid==2.9.3
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.5.0
pluggy==1.0.0
py==1.11.0
pylint==2.12.2
pyparsing==3.0.7
pytest==7.0.0
python-dotenv==0.19.2
requests==2.27.1
toml==0.10.2
tomli==2.0.1
typing-extensions==4.0.1
urllib3==1.26.8
wrapt==1.13.3
